requirements.txt
fastapi
uvicorn
torch
transformers
pillow
jinja2
